Replaced many floats with doubles

+++ Syren.Server/Hubs/DistanceHub.cs
@@ -8,7 +8,7 @@ public class DistanceHub(IDistanceService distanceService) : Hub
 {
     private readonly IDistanceService _distanceService = distanceService;
 
-    public void SendDistance(uint speakerId, float distance)
+    public void SendDistance(uint speakerId, double distance)
     {
         DistanceData distanceData = new() { SpeakerId = speakerId, Distance = distance };
         _distanceService.UpdateDistances([distanceData]);
