+++ Syren.Server/Models/DistanceData.cs
@@ -3,5 +3,5 @@ namespace Syren.Server.Models;
 public readonly struct DistanceData
 {
     public required uint SpeakerId { get; init; }
-    public required float Distance { get; init; }
+    public required double Distance { get; init; }
 }
